Extract periodic thread into own file
